requirements.txt
asgiref
dj-database-url
Django
gunicorn
pathlib
Pillow
psycopg2-binary
sqlparse
typing_extensions
whitenoise
django-ckeditor
